refactor: load Mux video client-only with a poster skeleton
The Mux player reads Date.now() at render, which can't run during prerender
under Cache Components. Load it via dynamic(ssr: false) wrapped in <Suspense>
with a skeleton that renders the Mux poster (the video's first frame, so no
layout shift). This makes <Video> self-contained, so showcase/[slug] no longer
needs a page-level Suspense (its only prerender blocker was the gallery video).

- video-mux.tsx: the Mux renderer, lazy-loaded client-only
- video.tsx: dynamic(ssr:false) + Suspense + poster VideoSkeleton
- showcase/[slug]: drop the now-redundant page-level Suspense

Verified: build green (89/89), Mux player absent from prerendered HTML (no
hazard), poster in the static shell, videos load + play client-side.
