+++ src/app/(site)/(pages)/post/[slug]/page.tsx
@@ -106,8 +106,8 @@ export async function generateStaticParams() {
   return slugs.map((slug) => ({ slug }))
 }
 
-// Streams the post body so client embeds (video/tweet) that read unstable
-// values (Date.now, runtime data) don't block the route from prerendering.
+// Streams the post body so the tweet embed's request-time data doesn't block
+// the route from prerendering.
 export default function Page({ params }: ProjectPostProps) {
   return (
     <Suspense fallback={null}>
